feat: so much work
Process documentation covers all of this, but the speed of using cursor to help build this means I got caught in a whirlwind of activity. As this continues it may be necessary to take small breaks within a day to talk about what is hapenning since things are hapennign at a great speed than they normally do.